forget about  diagonally adjacent

+++ Assets/ECS/GridIndex.cs
@@ -4,6 +4,6 @@ namespace GameOfLife.ECS
 {
     public struct GridIndex : IComponentData
     {
-        public ulong index;
+        public int index;
     }
 }
