+++ Assets/ECS/LeftNeighboorChunkInGrid.cs
@@ -4,18 +4,18 @@ namespace GameOfLife.ECS
 {
     public struct LeftNeighboorChunkInGrid : IComponentData
     {
-        public ulong Value;
+        public int Value;
     }
     public struct RightNeighboorChunkInGrid : IComponentData
     {
-        public ulong Value;
+        public int Value;
     }
     public struct UpNeighboorChunkInGrid : IComponentData
     {
-        public ulong Value;
+        public int Value;
     }
     public struct DownNeighboorChunkInGrid : IComponentData
     {
-        public ulong Value;
+        public int Value;
     }
 }
